+++ packages/actor-scraper/puppeteer-scraper/src/internals/crawler_setup.ts
@@ -191,7 +191,6 @@ export class CrawlerSetup implements CrawlerSetupOptions {
             proxyConfiguration: await Actor.createProxyConfiguration(this.input.proxyConfiguration),
             launchContext: {
                 useChrome: this.input.useChrome,
-                stealth: this.input.useStealth,
                 launchOptions: {
                     ignoreHTTPSErrors: this.input.ignoreSslErrors,
                     defaultViewport: DEFAULT_VIEWPORT,