+++ test/e2e/test-web-initial-cookies/test.mjs
@@ -55,7 +55,6 @@ await run(import.meta.url, 'web-scraper', {
         },
     ],
     useChrome: false,
-    useStealth: false,
     ignoreSslErrors: false,
     ignoreCorsAndCsp: false,
     downloadMedia: true,