+++ CHANGELOG.md
@@ -17,6 +17,7 @@
   - `cpuSnapshotIntervalSecs` and `memorySnapshotIntervalSecs` has been replaced with top level `systemInfoIntervalMillis` configuration
   - `maxUsedCpuRatio` has been moved to the top level configuration
   - `ProxyConfiguration.newUrlFunction` can be async. `.newUrl()` and `.newProxyInfo()` now return promises.
+- stealth mode has been removed in favour of fingerprints and fingerprints are now enabled by default
 
 2.3.0 / 2022/04/07
 ====================